Add support for continous scrolling when printing lines to the screen. If the screen fills up, it will automatically scroll when the next line is printed.
Refactored management for the two screen buffers.  They are now
continously moving based on a global offset that is managed by
the scroll position.  This makes the code simpler and supports the
continous scrolling feature.

Fix a bug where the proportional fonts where using random memory reads
as pixel data when drawing spaces.