Added initial animation support for frames composed of multiple bitmaps. Replaced several large GUI bitmaps with frames.  This saves a considerable amount of SPI flash in the media sector.
Fixed 2 bugs in oledBitmapDraw() where the bitmap would not display
correctly if the x position was not on a pixel word boundary.

Animations will eventually be media files in flash.  For now
they are in RAM and bitmap media file indices of 0x80 and above
are redirected to use the animation frames..